+++ src/components/AlphabetFilter/AlphabetFilter.tsx
@@ -21,7 +21,7 @@ export interface AlphabetFilterProps {
   /** Value used to set the text for a `Text` component that will be displayed above the letter buttons. */
   descriptionText?: string | JSX.Element;
   /** Value used to set the text for a `Heading` component. */
-  headingText?: string;
+  headingText?: string | JSX.Element;
   /** ID that other components can cross reference for accessibility purposes. */
   id?: string;
   /** Adds the `disabled` prop to the AlphabetFilter when true. */