+++ src/components/List/listChangelogData.ts
@@ -18,20 +18,4 @@ export const changelogData: ChangelogData[] = [
       "Updated the `title` prop to allow JSX to render custom heading elements for accessible heading hierarchy.",
     ],
   },
-  {
-    date: "2023-10-26",
-    version: "2.1.1",
-    type: "Update",
-    affects: ["Accessibility"],
-    notes: [
-      "Updates `tabindex` value from 0 to -1. See Accessibility section for details.",
-    ],
-  },
-  {
-    date: "2023-10-18",
-    version: "2.1.0",
-    type: "New Feature",
-    affects: ["Accessibility", "Documentation", "Functionality", "Styles"],
-    notes: ["Adds the NewsletterSignup component to the DS library"],
-  },
 ];