+++ src/components/SearchBar/SearchBar.tsx
@@ -55,7 +55,7 @@ export interface SearchBarProps {
   /** Optional string for the SearchBar's description above the component. */
   descriptionText?: string;
   /** Optional string for the SearchBar's heading text above the component. */
-  headingText?: string;
+  headingText?: string | JSX.Element;
   /** The text to display below the form in a `HelperErrorText` component. */
   helperText?: HelperErrorTextType;
   /** ID that other components can cross reference for accessibility purposes */