+++ src/components/VideoPlayer/VideoPlayer.tsx
@@ -26,7 +26,7 @@ export interface VideoPlayerProps {
    * and `videoType` props */
   embedCode?: string;
   /** Optional string to set the text for a `Heading` component */
-  headingText?: string;
+  headingText?: string | JSX.Element;
   /** Optional string to set the text for a `HelperErrorText` component */
   helperText?: HelperErrorTextType;
   /** ID that other components can cross reference for accessibility purposes */